refactor: NetworkPanel을 항목·탭 컴포넌트로 분리, 매치 항목 접기 가능
407줄짜리 NetworkPanel을 분리:
- NetworkListItem: 항목 1개 렌더(헤더·path·탭·콘텐츠)
- NetworkGeneralSection: General 탭(Request/Response와 대칭)
- utils/networkFormat: 색상·포맷 순수 함수

함께 매치 버그 수정: 매치로 자동 펼쳐진 항목이 탭해도 접히지 않던 문제를
collapsedIds 집합으로 해결(검색 외 단일 아코디언 동작은 유지).
