+++ package/src/components/NetworkPanel.tsx
@@ -1,14 +1,9 @@
 import { useState } from "@lynx-js/react";
-import {
-  matchNode,
-  type NetworkTab,
-  useNetworkSearch,
-} from "../hooks/useNetworkSearch";
+import { useNetworkSearch } from "../hooks/useNetworkSearch";
 import { useThemeColors } from "../styles/ThemeContext";
-import { fontWeight, type ThemeColors } from "../styles/theme";
+import { fontWeight } from "../styles/theme";
 import type { NetworkEntry } from "../types";
-import { HighlightText } from "./HighlightText";
-import { NetworkDetailSection } from "./NetworkDetailSection";
+import { NetworkListItem } from "./NetworkListItem";
 import "./NetworkPanel.css";
 import { NetworkSearchBar } from "./NetworkSearchBar";
 
@@ -17,133 +12,39 @@ interface NetworkPanelProps {
   clearNetworks: () => void;
 }
 
-const TABS: { tab: NetworkTab; label: string }[] = [
-  { tab: "general", label: "General" },
-  { tab: "request", label: "Request" },
-  { tab: "response", label: "Response" },
-];
-
-function getMethodColors(colors: ThemeColors, method: string) {
-  switch (method) {
-    case "GET":
-      return {
-        color: colors.palette.blue600,
-        backgroundColor: colors.palette.blue100,
-      };
-    case "POST":
-      return {
-        color: colors.palette.green600,
-        backgroundColor: colors.palette.green100,
-      };
-    case "PUT":
-      return {
-        color: colors.palette.yellow600,
-        backgroundColor: colors.palette.yellow100,
-      };
-    case "PATCH":
-      return {
-        color: colors.palette.purple600,
-        backgroundColor: colors.palette.purple100,
-      };
-    case "DELETE":
-      return {
-        color: colors.palette.red600,
-        backgroundColor: colors.palette.red100,
-      };
-    default:
-      return {
-        color: colors.fg.neutral,
-        backgroundColor: colors.bg.neutralWeak,
-      };
-  }
-}
-
-function getStatusCodeColor(
-  colors: ThemeColors,
-  variant: "success" | "error" | "pending",
-): string {
-  switch (variant) {
-    case "success":
-      return colors.palette.green600;
-    case "error":
-      return colors.palette.red600;
-    case "pending":
-      return colors.fg.neutralSubtle;
-  }
-}
-
-function getItemBg(colors: ThemeColors, status: string): string | undefined {
-  switch (status) {
-    case "pending":
-      return colors.palette.gray100;
-    case "error":
-      return colors.palette.red100;
-    default:
-      return undefined;
-  }
-}
-
 export const NetworkPanel = ({
   networks,
   clearNetworks,
 }: NetworkPanelProps) => {
   const colors = useThemeColors();
+  // 수동으로 펼친 단일 항목(아코디언). 검색 매치는 별도로 자동 펼침된다.
   const [selectedId, setSelectedId] = useState<string | null>(null);
+  // 매치로 자동 펼쳐졌지만 사용자가 직접 접은 항목들(자동 펼침을 무시)
+  const [collapsedIds, setCollapsedIds] = useState<Set<string>>(new Set());
   const search = useNetworkSearch(networks);
 
-  const isExpanded = (id: string): boolean =>
-    selectedId === id || search.isMatched(id);
-
-  const toggleExpanded = (id: string) =>
-    setSelectedId(isExpanded(id) ? null : id);
-
-  const formatDuration = (duration?: number): string => {
-    if (!duration) return "-";
-    if (duration < 1000) return `${duration}ms`;
-    return `${(duration / 1000).toFixed(2)}s`;
+  const isExpanded = (id: string): boolean => {
+    if (collapsedIds.has(id)) return false;
+    return selectedId === id || search.isMatched(id);
   };
 
-  const extractPath = (url: string): string => {
-    const pathMatch = url.match(/^https?:\/\/[^/]+(.*)$/);
-    if (pathMatch?.[1]) {
-      return pathMatch[1].startsWith("/")
-        ? pathMatch[1].slice(1)
-        : pathMatch[1];
+  const toggleExpanded = (id: string): void => {
+    if (isExpanded(id)) {
+      // 접기: 아코디언 선택을 해제하고, 매치 항목이면 자동 펼침도 끈다
+      setSelectedId((cur) => (cur === id ? null : cur));
+      if (search.isMatched(id)) {
+        setCollapsedIds((prev) => new Set(prev).add(id));
+      }
+    } else {
+      // 펼치기: 아코디언 선택 + 접힘 표시 해제
+      setSelectedId(id);
+      setCollapsedIds((prev) => {
+        if (!prev.has(id)) return prev;
+        const next = new Set(prev);
+        next.delete(id);
+        return next;
+      });
     }
-    return url;
-  };
-
-  const getGeneralInfo = (network: NetworkEntry) => {
-    return [
-      { key: "URL", value: network.url },
-      { key: "Method", value: network.method },
-      network.statusCode
-        ? { key: "Status", value: String(network.statusCode) }
-        : null,
-      {
-        key: "Request Time",
-        value: new Date(network.startTime).toISOString(),
-      },
-      network.endTime
-        ? {
-            key: "Response Time",
-            value: new Date(network.endTime).toISOString(),
-          }
-        : null,
-      network.duration
-        ? { key: "Duration", value: formatDuration(network.duration) }
-        : null,
-    ].filter((item) => item !== null);
-  };
-
-  const getStatusCodeVariant = (
-    status: string,
-    statusCode?: number,
-  ): "success" | "error" | "pending" => {
-    if (status === "pending") return "pending";
-    if (status === "error") return "error";
-    if (statusCode && statusCode >= 200 && statusCode < 300) return "success";
-    return "error";
   };
 
   return (
@@ -190,216 +91,22 @@ export const NetworkPanel = ({
           scroll-orientation="vertical"
           className={"np-list"}
         >
-          {networks.map((network) => {
-            const activeTab = search.getActiveTab(network.id);
-            return (
-              <list-item key={network.id} item-key={network.id}>
-                <view
-                  className={"np-item"}
-                  style={{
-                    backgroundColor: getItemBg(colors, network.status),
-                    borderBottomColor: colors.stroke.neutralWeak,
-                  }}
-                >
-                  <view
-                    className={"np-itemHeader"}
-                    bindtap={() => toggleExpanded(network.id)}
-                  >
-                    <text
-                      className={"np-method t2"}
-                      style={{
-                        fontWeight: fontWeight.bold,
-                        ...getMethodColors(colors, network.method),
-                      }}
-                    >
-                      {network.method}
-                    </text>
-                    {network.statusCode && (
-                      <text
-                        className={"np-statusCode t2"}
-                        style={{
-                          fontWeight: fontWeight.bold,
-                          color: getStatusCodeColor(
-                            colors,
-                            getStatusCodeVariant(
-                              network.status,
-                              network.statusCode,
-                            ),
-                          ),
-                        }}
-                      >
-                        {network.statusCode}
-                      </text>
-                    )}
-                    {network.status === "pending" && (
-                      <text
-                        className={"np-statusCode t2"}
-                        style={{
-                          fontWeight: fontWeight.bold,
-                          color: colors.fg.neutralSubtle,
-                        }}
-                      >
-                        Pending...
-                      </text>
-                    )}
-                    <text
-                      className={"np-time t2"}
-                      style={{
-                        fontWeight: fontWeight.regular,
-                        color: colors.fg.neutralSubtle,
-                      }}
-                    >
-                      {formatDuration(network.duration)}
-                    </text>
-                    <text
-                      className={"np-time t2"}
-                      style={{
-                        fontWeight: fontWeight.regular,
-                        color: colors.fg.neutralSubtle,
-                      }}
-                    >
-                      {new Date(network.startTime).toISOString()}
-                    </text>
-                  </view>
-
-                  <view bindtap={() => toggleExpanded(network.id)}>
-                    <HighlightText
-                      text={extractPath(network.url)}
-                      query={search.searchQuery}
-                      className={"np-path t3"}
-                      style={{
-                        fontWeight: fontWeight.regular,
-                        color: colors.fg.neutral,
-                      }}
-                    />
-                  </view>
-
-                  {isExpanded(network.id) && (
-                    <view
-                      className={"np-detailsContainer"}
-                      style={{ borderTopColor: colors.stroke.neutralSubtle }}
-                    >
-                      {/* Tabs */}
-                      <view className={"np-tabs"}>
-                        {TABS.map(({ tab, label }) => {
-                          const isActive = activeTab === tab;
-                          return (
-                            <view
-                              key={tab}
-                              className={"np-tab"}
-                              style={{
-                                backgroundColor: isActive
-                                  ? colors.bg.neutralWeak
-                                  : undefined,
-                              }}
-                              bindtap={() => search.selectTab(network.id, tab)}
-                            >
-                              <text
-                                className={"np-tabText t4"}
-                                style={{
-                                  fontWeight: fontWeight.medium,
-                                  color: isActive
-                                    ? colors.fg.neutral
-                                    : colors.fg.neutralSubtle,
-                                }}
-                              >
-                                {label}
-                              </text>
-                            </view>
-                          );
-                        })}
-                      </view>
-
-                      {/* Tab Content */}
-                      <view className={"np-tabContent"}>
-                        {activeTab === "general" && (
-                          <view className={"np-table"}>
-                            {getGeneralInfo(network).map((item) => (
-                              <view
-                                key={item.key}
-                                ref={
-                                  item.key === "URL"
-                                    ? search.getNodeRef(
-                                        network.id,
-                                        matchNode.url,
-                                      )
-                                    : undefined
-                                }
-                                className={"np-tableRow"}
-                                style={{
-                                  backgroundColor: colors.bg.neutralWeak,
-                                }}
-                              >
-                                <text
-                                  className={"np-tableKey t3"}
-                                  style={{
-                                    fontWeight: fontWeight.bold,
-                                    color: colors.fg.neutralSubtle,
-                                  }}
-                                >
-                                  {item.key}
-                                </text>
-                                <HighlightText
-                                  text={item.value}
-                                  query={
-                                    item.key === "URL" ? search.searchQuery : ""
-                                  }
-                                  activeOccurrence={
-                                    item.key === "URL"
-                                      ? search.getActiveOccurrence(
-                                          network.id,
-                                          matchNode.url,
-                                        )
-                                      : -1
-                                  }
-                                  className={"np-tableValue t3"}
-                                  style={{
-                                    fontWeight: fontWeight.regular,
-                                    color: colors.fg.neutral,
-                                  }}
-                                />
-                              </view>
-                            ))}
-                          </view>
-                        )}
-
-                        {activeTab === "request" && (
-                          <NetworkDetailSection
-                            section="request"
-                            headers={network.requestHeaders}
-                            body={network.requestBody}
-                            highlightQuery={search.searchQuery}
-                            getActiveOccurrence={(nodeKey) =>
-                              search.getActiveOccurrence(network.id, nodeKey)
-                            }
-                            getNodeRef={(nodeKey) =>
-                              search.getNodeRef(network.id, nodeKey)
-                            }
-                          />
-                        )}
-
-                        {activeTab === "response" && (
-                          <NetworkDetailSection
-                            section="response"
-                            headers={network.responseHeaders}
-                            body={network.responseBody}
-                            error={network.error}
-                            highlightQuery={search.searchQuery}
-                            getActiveOccurrence={(nodeKey) =>
-                              search.getActiveOccurrence(network.id, nodeKey)
-                            }
-                            getNodeRef={(nodeKey) =>
-                              search.getNodeRef(network.id, nodeKey)
-                            }
-                          />
-                        )}
-                      </view>
-                    </view>
-                  )}
-                </view>
-              </list-item>
-            );
-          })}
+          {networks.map((network) => (
+            <list-item key={network.id} item-key={network.id}>
+              <NetworkListItem
+                network={network}
+                expanded={isExpanded(network.id)}
+                onToggle={() => toggleExpanded(network.id)}
+                activeTab={search.getActiveTab(network.id)}
+                searchQuery={search.searchQuery}
+                onSelectTab={(tab) => search.selectTab(network.id, tab)}
+                getActiveOccurrence={(nodeKey) =>
+                  search.getActiveOccurrence(network.id, nodeKey)
+                }
+                getNodeRef={(nodeKey) => search.getNodeRef(network.id, nodeKey)}
+              />
+            </list-item>
+          ))}
         </list>
       )}
     </view>
